feat(cli): add-type wizard talking face greeting (ITERATION-339)
Reuses the say/should_greet pattern from init's greeting so the
add-type wizard gets the same talking-face treatment, gated the same
way behind TTY/--json/no-colour. Also consolidates has_pollable_types
into tui.rs so app.rs and views.rs share one gate for the sync
indicator and background poll instead of duplicating the store check.
