+++ src/tui/infra/event_loop.rs
@@ -235,18 +235,6 @@ fn try_push_clickup_edit_with<C: ClickupClient + 'static>(
         .map_err(|e| e.to_string())
 }
 
-// Whether the background poll should run for this project: true when any type is
-// backed by a store the poll refreshes (github issues/milestones OR clickup
-// tasks). Milestone-only and clickup-only projects still need the poll so a
-// milestone/task created after launch appears live in the list.
-fn has_pollable_types(config: &Config) -> bool {
-    config.documents.types.iter().any(|t| {
-        t.store == StoreBackend::GithubIssues
-            || t.store == StoreBackend::GithubMilestones
-            || t.store == StoreBackend::ClickupTasks
-    })
-}
-
 // Human-readable summary of a fix run for the warnings panel. The engine op core
 // returns the structured `FixOutput`; each frontend renders its own
 // presentation, so this mirrors the CLI's `fix::output::format_human`
@@ -684,24 +672,25 @@ pub fn run(store: Store, config: &Config) -> Result<()> {
     let (tx, rx) = crossbeam_channel::unbounded();
     app.event_tx = tx.clone();
 
-    let shared_gh_store: Option<Arc<Mutex<GithubIssuesStore>>> = if has_pollable_types(&config) {
-        let gh_config = config.documents.github.as_ref();
-        let repo = gh_config.and_then(|g| g.repo.clone());
-        repo.map(|repo| {
-            let root = app.store.root();
-            Arc::new(Mutex::new(GithubIssuesStore {
-                client: Box::new(GhCli::new()),
-                root: root.to_path_buf(),
-                repo,
-                config: config.clone(),
-                issue_map: IssueMap::load(root)
-                    .unwrap_or_else(|_| serde_json::from_str("{}").unwrap()),
-                issue_cache: IssueCache::new(root),
-            }))
-        })
-    } else {
-        None
-    };
+    let shared_gh_store: Option<Arc<Mutex<GithubIssuesStore>>> =
+        if crate::tui::has_pollable_types(&config) {
+            let gh_config = config.documents.github.as_ref();
+            let repo = gh_config.and_then(|g| g.repo.clone());
+            repo.map(|repo| {
+                let root = app.store.root();
+                Arc::new(Mutex::new(GithubIssuesStore {
+                    client: Box::new(GhCli::new()),
+                    root: root.to_path_buf(),
+                    repo,
+                    config: config.clone(),
+                    issue_map: IssueMap::load(root)
+                        .unwrap_or_else(|_| serde_json::from_str("{}").unwrap()),
+                    issue_cache: IssueCache::new(root),
+                }))
+            })
+        } else {
+            None
+        };
 
     let cache_ttl = config
         .documents
@@ -712,7 +701,7 @@ pub fn run(store: Store, config: &Config) -> Result<()> {
     // Schedule polling whenever the project has any pollable type, independent of
     // whether a github store was built: a clickup-only project has no github repo
     // (shared_gh_store == None) but must still poll to refresh its task cache.
-    let mut next_poll = if has_pollable_types(&config) {
+    let mut next_poll = if crate::tui::has_pollable_types(&config) {
         Some(Instant::now())
     } else {
         None
@@ -1077,41 +1066,6 @@ mod tests {
     use super::*;
     use crate::engine::config::TypeDef;
     use tempfile::TempDir;
-
-    // Gate: a project whose only GitHub-backed type is github-milestones must
-    // still poll, so a milestone created after launch appears live in the list.
-    #[test]
-    fn milestone_only_project_is_pollable() {
-        let mut config = Config::default();
-        config.documents.types = vec![TypeDef::test_fixture(
-            "milestone",
-            StoreBackend::GithubMilestones,
-        )];
-
-        assert!(has_pollable_types(&config));
-    }
-
-    // Gate: a clickup-only project must poll too, so tasks created after launch
-    // appear live without a manual fetch — same parity as github types.
-    #[test]
-    fn clickup_only_project_is_pollable() {
-        let mut config = Config::default();
-        config.documents.types = vec![TypeDef::test_fixture("task", StoreBackend::ClickupTasks)];
-
-        assert!(has_pollable_types(&config));
-    }
-
-    // Gate: a project with no pollable types must not poll.
-    #[test]
-    fn project_without_gh_types_is_not_pollable() {
-        let mut config = Config::default();
-        config.documents.types = vec![
-            TypeDef::test_fixture("doc", StoreBackend::Filesystem),
-            TypeDef::test_fixture("note", StoreBackend::Filesystem),
-        ];
-
-        assert!(!has_pollable_types(&config));
-    }
 
     // Build an App over `root` with the given config, using a deterministic
     // halfblocks picker so no terminal probing happens in tests.
